+++ Sphere.cpp
@@ -6,26 +6,26 @@ Sphere::Sphere(){
 
 }
 
-Sphere::Sphere(vec3 c, float r){
+Sphere::Sphere(vec3 c, double r){
     center = c;
     radius = r;
 }
 
-bool Sphere::hit(const Ray& r, float t_min, float t_max, hit_record& rec) const{
+bool Sphere::hit(const Ray& r, double t_min, double t_max, hit_record& rec) const{
     // get the ray to the center of the sphere
     vec3 origin_to_center = r.origin() - center;
 
     // get the numbers needed to solve for the ray intersection.
     // check for solutions using quadratic formula
-    float a = dot(r.direction(), r.direction());
-    float b = 2 * dot(origin_to_center, r.direction());
-    float c = dot(origin_to_center, origin_to_center) - radius * radius;
+    double a = dot(r.direction(), r.direction());
+    double b = 2 * dot(origin_to_center, r.direction());
+    double c = dot(origin_to_center, origin_to_center) - radius * radius;
 
-    float discriminant = b*b - 4*a*c;
+    double discriminant = b*b - 4*a*c;
 
     // first, check if a solution exists
     if(discriminant > 0){
-        float t = (-b - sqrt(discriminant)) / (2 * a);
+        double t = (-b - sqrt(discriminant)) / (2 * a);
         
         // check if the solution is within our bounds...
         if(t_min < t && t < t_max){
